Add a video module for vimeo.com
The trick is to log ourselves in to validate the provided signature,
then the video file redirect works.
Since the redirected url is not checked for User-Agent unlike the redirect one,
we disable redirects to get the actual location and use it instead.
This allows running wget or vlc without faking their User-Agents.

For now it only downloads sd videos.
TODO: get the highest quality
TODO: search, latest

Signed-off-by: François Revol <revol@free.fr>
Signed-off-by: Romain Bignon <romain@symlink.me>