gdcvault: Fix various weird cases
* comment out or remove some debug prints
* try to cope with non-UTF-8 title passed (1013483).
Couldn't get it to be recoded yet.
* handle boggus url to xml config file (http:/...) (1015020)
* handle iframes with page name different than 'player.html' (1013798)
* handle xml config filenames with spaces (1441)
* catch xml config names with 'smil' as hostname, which means speakervideo
points to a smil file describing the streams, and take the file with the
highest bitrate
* account for xml config with only valid slidesvideo (1016627) and use it
as fallback
* handle configs with 'Invalid Date' as date text (1016634)

We can now dump the entire gdcvault video files urls with a few exceptions
(403 HTTP errors), and many /mediaProxy.php urls failing to redirect, which
are still to be investigated.