add ICapImage and BaseImage
BaseImage was extracted from the gallery capability and
refactored to inherit from BaseFile.
As I did not look further into the gallery capability, I left
the original BaseImage there.
I'm aware that it's not an healthy situation, but it seems
to be used only by modules using ICapGallery and hopefully
should not have any consequences for now

Signed-off-by: Pierre Mazière <pierre.maziere@gmx.com>