Fix bug dailymotion not playing video with "Server returned 403: Forbidden"
Dailymotion video url send cookies then use a 302 redirect to the real video url
The real video url check the cookies sent in the request.

Some mediaplayer (mplayer) can not handle this.

So we use pure python urllib2 to handle this and load the video.
The video bytes are piped to the mediaplayer.

Others modules might need this.
They just have to set the "_play_proxy" boolean attribute to the video object.

Signed-off-by: Vincent Texier <vit@free.fr>