Use a fast path for ls if -U is set
The ls command first store all objects and collections, and then run the
formatting. It is bad for applications displaying results as soon as
possible, but mandatory to store results and display the collections
first.

Since the -U option was introduced to not sort the results, we can
expand this idea to not display the collections first, and to display
results immediately.