pastebin: Convert to browser2, fix some little things
Stop testing public posts, this has been failing for a long time.
Retrieve the date, though it isn't in the capability yet.
Move most of the code in browser.py now that it is cleaner and simpler.

Some more improvements should be coming, but it is already in a much
better state than it was.