Add a DebugFilter class to weboob/tools/log.py
The browser2 filter debugging is an extra level of "Debug". Since python
logging does not support adding severity levels, we use the Filter class
to remove the lines by default.

Sadly, we cannot pass variable to filters (logging system does not use
the instance passed with addFilter method), so we use to write a filter
for each use case (today: only one...)