weboobcfg: Format capabilities as JSON array
Capabilities was always displayed as a single list of entries
separated by comma. This is not really efficient as the reader has to
parse this string to retrieve atomic informations, even in JSON.

This patch allows to display the capabilities as an array, allowing
standard JSON parsers to immediatly returns a valuable information.