ing: fix bad login with cookie cache
Description of the problem: we are loading an expired cookie. We go to
the login page (redirected by the website, or because of the url stored
in storage. It does not matter).
Next step: we call do_logout, it clear the cookies
Next step: stay_or_go do nothing
Next step: we do a post, without any cookie. The request fail

Solution: force the "go()" after cookie cleaning