fix bug when card is first account
Previously card accounts were linked to the previous account, but we
encountered a card account in first position. We decided to link it to
following account because doing something cleaner was to demanding. We
think this should be enough because this behaviour is rare and an even
more complex one should not exist.