bp: Fix for new API
As of November 2013, there were a few changes made to the PDFMiner API
prior to October 2013. This is the result of code restructuring. Here is
a list of the changes:
PDFDocument class is moved to pdfdocument.py.
PDFDocument class now takes a PDFParser object as an argument.
PDFDocument.set_parser() and PDFParser.set_document() is removed.
PDFPage class is moved to pdfpage.py
process_pdf function is implemented as a class method PDFPage.get_pages.