qt5/models: add ResultModel for viewing objects hierarchies and FilterTypeModel
ResultModel is a Qt model to show weboob objects suitable for standard
QTreeViews and QListViews. Rows representing containers (like Collection or
Gallery) can be expanded to fetch their sub-objects.
Fields can be associated to columns. Thumbnails are displayed.

FilterTypeModel is a proxy model for filtering a ResultModel basing on the
object types, for example in a QTreeView to show only Collections.