browser: introduce PARENT_ATTR class attribute
When the module does not have a BROWSER attribute (for example, modules
with several browser) we cannot use the AbstractBrowser: we got one
horrible exception since module.klass.BROWSER is None.

This patch add the PARENT_ATTR option, to get any class loaded somewhere
in the parent module. It also add a better exception message when we
fail to load the parent browser