Make as much modules unittests as possible run without backend
Using the `nofail` argument and hooking to `login` event, one can make
as much modules unittests as possible run, and skip the unittests which
require authentication or special configuration.

Tests are also run using all available backends now. This is especially
important for modules such as Amazon for which the backend has a large
influence on the website being used and the code being run.

I also changed the default values so that they match the expected type
(rather than being `None`).