Fix date translation
This does not works as intended:

	re.sub(u'\bdéc\b', 'december', '31 déc, 21h02')

It is because u'\b' is passed to the regexp engine as '\x08' (which is
NOT the word boundary location match, it just matches the ASCII backspace
non-printable character). The correct way to pass a \b to the regexp
engine is either u'\\b' or ur'\b'.

Signed-off-by: Simon Lipp <laiquo@hwold.net>