weboob.capabilities.base.BaseObject: copy() should also copy Field objects
When calling copy() method which does a shallow copy, the _fields dict
was copied, but not its values (which are Field instances). Since the
real content of the BaseObject is stored in the Field instances, the
values were shared instead of being shallow-copied:

    obj2 = obj1.copy()
    obj2.foo = 'bar'  # would modify obj1.foo

Since _fields is kind of read-only, it's useless not to copy the Field
instances.