elements: better handling of SkipItem
We sometimes raises a SkipItem exception in attributes parsing, instead
to use filters like "condition" of "validate" methods. Even it was not
designed for that before, it's sometimes useful and easier to do so.

Since we accept it as a valid behavior (upper levels handle this
exception), we should not flood our users with a warning