[bforbank] parse when multiple deferred cards one same checking account
The "index" of the card can only be found in an history page, and same
for the balance.

For the summary transactions, we used to search them in the checking
account and dispatch to the card account, but with multiple cards, it
doesn't fit, so we generate a transaction directly.