weboob.browser.browsers.AbstractBrowser: don't create a new type dynamically
Creating a new type on-the-fly prevents from using isinstance, super and
more.
We can get rid of this by changing the class hierarchy dynamically.

Also, the modules should now pass the Ouiboube instance as a
keyword-argument to the Browser, so the arguments do not have to be
modified between __new__ and __init__.