qgalleroob: results can be fetched on-demand, and can be permanently hidden
Some modules can return a lot of items. Instead of consuming all results
immediately, results can now be fetched on demand. That avoids having to
set a low limit to number of results.
This is done by not closing the iterator iter_gallery_images etc.
The app uses a Queue to pause/resume iteration.

Also, sometimes it's useful to do the same search, but some results are
unwanted, so we may want to hide them permanently, to avoid fetching their
data and thumbnail next time the iterator reaches them.
Items can be bookmarked or ignored.