tools: only run python2 or python3, not both
The solution to check if a version is installed is not enough for CI. We
often have python2 installed, but not onfigured.

This patch stops the behaviour to run both versions, a caller must now
call the script twice to check both pythons versions. It's not perfect,
but probably easier to understand