[bnporc] merge list of cards
At the end of the month, the list of cards is duplicated,
because the coming amount may exist for the current month
and the next. We need to merge the list of cards in this
situation, we try to only keep cards with a coming amount
in case of a duplicate.