boobank: make "transfer" command more interactive and add "recipients" command
The "transfer" command could do a transfer and also list recipients.
Split its behavior by introducing the "recipients" command. It uses an
appropriate Recipient formatter.

The "transfer" command is only for making transfers. If all params are
passed, the transfer is performed. If some params are missing and
boobank is interactive, they will be queried interactively.