cmb/cmmc: replicate cmso BackendConfig
cmb and cmso are AbstractModules of cmso, which allows them to inherit cmso's
content. However, the inheritance is only done when instanciating the module.
When adding a backend with weboob command-lines, the module is NOT
instanciated yet and thus does not inherit the CONFIG attribute, so the
default empty CONFIG is used and the module cannot be configured.

https://git.weboob.org/weboob/devel/issues/44

A workaround is to copy the CONFIG of the PARENT module so it does not need
to be instanciated.

A real fix would be to find a way to inherit the PARENT module class before
using the module's CONFIG (or anything else), with a dependency system at
best.