weboob.tools.value: add an "aliases" parameter to handle old "choices"
For the "banquepopulaire" module, the "website" config should change
because the list has changed on website. Unfortunately, the old choices
continue to exist in config files so they must still be accepted by the
module, and thus they cannot be removed from the choices list, even
though they are irrelevant for new users.

Add an optional "aliases" param along with the "choices" to Value, to
have a list of accepted compatibility aliases, and still present an
up-to-date choices list.