[myedenred] Remove ids after getting doublons
Here is the fact. Before, we were creating unique ID from date, row, and others informations that were disponible in the website. Transaction's id aren't mandatory, and it was pretty ugly. Adding to that the fact that we recently got a case where there was two exactly similar transactions, causing a crash, we deceided to delete it.