[cragr] fixes the retrieval of an investment's vdate
When retrieving the vdate of an investment, the value sent to be filtered by Date()
was an object Element causing the filtering to fail (AttributeError: 'list' object has no attribute 'read').
By calling CleanText first, the value sent is the date in unicode. The right value, with no errors, is returned

Closes: 500983@redmine