[cragr] Coming on card accounts
With the new decision on card accounts, the balance has no place to be.
With this commit we will always have the balance at 0 and the operations
will be in the commit field.
We will skip the not deferred card who appears on the same page but has
no history
Closes : 4037@zendesk