[boursorama] ignore another format of opposed cards
Sometimes, opposed cards have a link and their history can be consulted.
However they are not displayed on cards page, only on accounts list.

They can be detected as being opposed by their label, but only on the
AccbisPage though. Their label on AccountsPage does not show if it's
opposed or not.