societegenerale: skip GDPR form and fix add recipients
On login, there is a form to ask user validating GDPR questions. It can
be skipped by calling a specific URL.

Also, the link to add a recipient was broken. To prevent eventual
compatibility break, the old XPath is kept.

NOTE: I CAN'T TEST REAL RECIPIENT ADD OR TRANSFERS, SO PERHAPS IT IS
STILL BROKEN LATER IN THE PROCESS.