[bforbank] handling transition period between last accounting day and debit date for deffered card accounts
Closes: 4157@zendesk

The problem came from bforbank site displaying next month page as soon
as the last accounting day is past, not the debit date. For example,
if the last accounting day is 26/02, and the debit date 30/02, and assuming
we are the 27/02, the site displays the page for the next month (with debit
date to the 30/03). As a result, balance was not correct during this period,
as for coming transactions.
To handle this, we check if the month of the debit date we scrap on the main
card history page is equal to the current month + 1 (we are the 27/02,
but because the last accounting day was 26/02, the main page debit date is 30/03).
If so, we need to go to the previous month page.