[cices] repaired investment
At first, there were FCPE and CCB for investments. Now the CCB is very rare and from what I saw, it does not look like an investment anymore.

So I redid the Investment only for the FCPE which does not take into account multpiple rows and kept the CCB, adapting it to the rest.