[orange] no-pro subscriptions download documents from the website
Some subscriptions used to be pro but even now that they are regular they are still build through the API.
However this way prevents from downloading the documents because the API does not allow it.
So using the id, the document is downloaded from the website.

And I removed a 'dayfirst' parameter because it messed up the id building.

Closes: 613018@redmine