[bouygues] refactor module to parse many subscriptions
Closes: 4695@zendesk 4589@zendesk

On this website there is a many to many relation between subscription and bill.
so we have to consider 'compte-facturation' as a subscription to get a one to many relation with bill
by the way it's still possible to have many 'compte-facturation',
but it's not a problem since there is no redundant bill among subscription.