[axabanque] access to investments
There were a redirection problem.

First, the BoursePage url/form changed.

Second, you have an iframe inside another one, but the first has no src. So a form is used to acces the second.
It leads you to a redirection page which I make redirect to another page. And on this page, you have the second iframe.

I had to modify lcl so it takes handle_refresh into account.

Closes: 4815@zendesk