add CapBank binck module
This module adds support for Binck stock trading accounts.
It can show the list of accounts with their current EUR valuation, the list of investments for each account in the correct currency, and it can show the list of transactions (but this list sucks on binck.fr to the point of being almost unusable, this backend cannot invent meaningful descriptions so it does its best).