[banquepopulaire] get references on gocardless transactions
To get the references, one must open the detail page for one particular transaction.
To do so, the module get the parameters, load the page and return to the transaction page.
Knowing that to get the details, you must know and retrieve the transaction id which changes on each loading of the transaction page.

The references are stored in the raw field.

Closes: 5152@zendesk