[CMSO] fixes navigation in order to be more consistent with regular browser one
Issue found : normal navigation supposes that we go back to account page before consulting
history. In a certain number of cases, the server will raise a 500 error because we didn't
follow regular navigation.
Fix : before going to an account's history we go back to account page and we update
the csrf parameter in the history link.