[cmes] repaired iter_pocket
At first the probleme was that the investments where not stored in the backend.
The reason was that the fetching of pockets crashed so nor the investments nor the pockets were saved.

ListElement is used instead of TableElement because the strange layout of the table shifts the columns header.

The strange thing, here, is that the boobank command works fine locally, just giving no pockets, without any error.
It was because the CapBank was not the good one. So now instead of 'CapBankWealth', it is 'CapBankPockets'.

cices and pradoepargne has been updated because depending on cmes.

Closes: 5313@zendesk