[societegenerale] parent for card account
There is two different layouts on the detail card page.
A list of the cards linked to one account, with the id oh this one.
Or the history of the card account with a button showing the details.
The details are already on the html page, the button is just used to display them.