[americanexpress]Scrapping all accounts in the new website
Now americanexpress have one JSON file per main account. First we only get the token from the main account, we takes the name of
the owner and finally we compares it with the embossed_name in the JSON. Nevertheless, old token were kept for the balance and the coming values.