[creditdunord] scraped 'COMPTE A TERME'
The 'COMPTE A TERME' are on another page. So just got to get there and iterate.

For the two accounts on the tested connection, they had the same id and the same label.
So an adding is made to distinguish them.

Moreover, they had no informations. Clicking on the link gives you a pop-up telling you there is no details.
So there is no history or coming.

Concern the 5295@zendesk