[barclays] Corrected coming and balance for DIFF cards
In this module the deferred credit cards appeared without any coming and their
coming was displayed in the balance column. Since the "coming" of each
card appear in the "balance" column on the BMilleis website, this has been fixed by
defining "obj_balance" and "obj_coming" methods that depend on the type of each account.

Closes: 5082@zendesk