[creditcooperatif] add nuser
Since caissedepargne has now a third identificator, nuser. It had been added to creditcooperatif.

With a Switching Browser, the module can use the CenetBrowser.

Several BASEURLs are handled because one is used for the login ans the other one for the rest of the navigation.
So that the BASEURLs can be written in the code for another script.

The migration script is the one made for caissedepargne to handle nuser.

Closes: 5588@zendesk
Closes: 5523@zendesk
Closes: 5574@zendesk
Closes: 5643@zendesk
Closes: 5645@zendesk
Closes: 4440@sibi