[caissedepargne] Add history with linebourse module
The history of some accounts are on the offrebourse website, which seems
to be the same than linebourse. So the patch allows the connection to
the site whith the help of linebourse module. Moreover the patch takes
care of the cookie with a deletion when CTX value appears twice (avoid a
403 connection).

Closes: 5556@zendesk