[cragr] Modified market balance update to avoid doublons for liquidities
The PEA balance was previously updated from the market website, however
this amount includes liquitidies that are already counted in the
associated DAV PEA account, leading to duplicates. For PEA accounts it is
therefore better to keep the amount scraped on the accounts list page.
The double 'for' loop forced boobank to connect to the market page
several times; using a separated method with its own double loop enables
account matching with only one connection to the market page.

Closes: 5621@zendesk