[banquepopulaire] Changed get_history return to prevent gocardless crash
The get_history method used a for loop a yielded transactions one by
one, which led to a crash if the gocardless information filling did not
work properly. I added a return in case we land on the ErrorPage.
This patch fills the transaction list while filling the gocardless information
for each gocardless transaction and eventually returns the transaction_list:
this way, the transaction list is always properly returned without crash due
to bugs in the set_gocardless_transaction_details() method.

Closes: 5879@zendesk