weboob.browser.nss: implement AIA support
Some sites rely on TLS extension "AIA" and do not provide a complete
certificate chain up to the CA. The AIA extension lets the site define
an HTTP URL where to fetch the parent certificates.

When encountering this case, the parent certificate must be checked to
really be the parent of the certificate being validated. Also, the
parent certificate must be in the trusted CAs.