[bp] Added "compte attente" as TYPE_CHECKING
"Compte attente" was of type_unknown, there is a current discussion
concerning switching the account types to an Enum system (see with
ntome) but for now we chose not to add a new type of account
(TYPE_ACCOUNTANCY). In the meantime, comptes attente will be typed as
type_checking, which is the closest type available in boobank for now.

Closes: 5436@zendesk