[bnporc] market account if only one account
In some case, you may have only one market type account. If so, the website does not show a page with the list of one account but directly the details of this account.

Most of the time, getting the number in the account label is enough to recognize it in the iter_accounts.
But sometime you have custom account label, so it is not enough. That is why the entire account name is retrieved.

Closes: 13125@sibi
Closes: 12518@sibi
Closes: 13576@sibi
Closes: 12610@sibi
Closes: 12871@sibi
Closes: 11832@sibi
Closes: 11912@sibi
Closes: 11913@sibi
Closes: 11867@sibi
Closes: 11750@sibi
Closes: 11749@sibi
Closes: 11580@sibi
Closes: 11715@sibi
Closes: 11714@sibi