[bnporc] get rid of bad history for enterprise
A recurring problem with bnporc is that sometimes you may not have the good transactions list.
It may be truncated a lot. For example, from 50 transactions, you had only 5.
Because of that, in the backend some transactions become deleted because of one update with the bad transactions.

I tried to fix it by copying the website, since you never had bad transactions on it.
First, I changed the 'typeReleve' from 'Comptable' to 'Previsionnel' because it is the only type the website is using.
And I added some parameters to the history request. The same ones are used to go on the transaction page and to get the json data.

I tested a lot on connections, knowing the right and the wrong history state.
Since I applied the patch, I had not once the wrong state.

Closes: 5663@zendesk
Closes: 6219@zendesk
Closes: 6225@zendesk
Closes: 6229@zendesk
Closes: 6115@zendesk