[bp] invalid account id
At first the account id was taken from an option tag. But this id was used to look for accounts that begin with this id.
Now it looks for the id in the account table.

This case conflicts with 1 table student loan which needs the id in the option tag.

Changing the order of the ifs is enough.

Closes: 15258@sibi
Closes: 15334@sibi
Closes: 15335@sibi
Closes: 6592@zendesk
Closes: 6334@zendesk