[spirica] Added symbols='-' for the transaction dates
A weboob contributor proposed a patch to fix the Date when empty:

"For the field date, the module use the column "Date de réception"
and "Date de valeur" if the first is empty. But Spirica uses "-" when
the cell is empty, so no date is returned and boobank uses today's date
as default value."

I adapted his merge request since we now use CleanText instead of
TableCell for iter_history, with the connection I have the symbols='-'
does not change any date but maybe in some cases it will prevent boobank
from returning today's date by default.