[creditcooperatif] duplicate transactions
Looking at the website, you can see that some transactions are duplicates.
So it is normal to have some, since it is what is send.

But sometimes, some requests send have the same transaction list, even if they had different parameters/data.
So there are some more duplicates.
It happens only for the card type transactions.

They are compared to be sure they does not bring duplicates.
It is pretty easy since their labels always have the same format: 'TOT DIF %MONTH%'.
And the duplicates one have the same amount too.
The rest is checked just to be sure.

Closes: 6257@zendesk