[creditmutuel/cic] Remove trash from Transactions labels
The infamous "Cliquer pour déplier ou plier le détail de l’opération" is
back!

This is linked to the commit 6d41bfaf that removes this piece of some
transaction labels ; but it turns out that now it's not located in
parts[0] instead of parts[1].
I first remove all empty strings and then look for the string that we
want to get rid of ; if it's there I remove it from the list, this way
it is more dynamic that parts[0] or parts[1].

Tested with CIC and creditmutuel connections.

Closes: 6378@zendesk, 6382@zendesk, 6381@zendesk, 6398@zendesk