weboob.capabilities: replace stdlib enum with custom enums
Python's enums are not extensible, values can't be added and they can't
be subclassed. This is problematic if weboob devel adds new values,
because stable weboob will not be able to use them. They would need to
be replaced by a default value or None, which loses information.

Instead, implement custom enums that contain plain constants and support
adding values aftewards.