[n26] wrongpass because of bad exception
A lot of wrongpass appeared on n26 connector.
It is because when tying to login, after a number of failures, you cannot log in right away.
You need to wait 30 minutes, that is what is said by a message.

Now the message is retrieved in a BrowserUnavailble exception.
The BrowserIncorectPassword also has an error message.

Closes: 6279@zendesk
Closes: 6520@zendesk