[banquepopulaire] wrong catching of the login page
When the user is already logged in and it tries to log anyway, the website response by a json page which says it.
But sometimes, this page may be catch by the Login2Page because of the URL.

So now, the page exist with a is_here to be sure.

Closes: 16413@sibi
Closes: 16408@sibi
Closes: 17916@sibi
Closes: 10663@sibi
Closes: 13746@sibi
Closes: 10622@sibi
Closes: 16474@sibi
Closes: 15877@sibi
Closes: 16473@sibi
Closes: 17884@sibi
Closes: 18531@sibi
Closes: 15851@sibi