[cmso] confirmed investments list and added vdate and code_type
There was an assert in the cmso/pro/browser.py code to verify that the
investements list appears correctly when there are more than one
investment. The two investments of the concerned connection appear
correctly in boobank so I deleted the assert.
In addition, I completed the liquidities, vdate, code and code_type information
for both PAR and PRO websites.