[ing] Fixed excessive requests to get transaction investment ISIN codes
The ING module was performing many requests for the transaction list
since it was performing an Async to get the ISIN code for each investment of each
transaction (up to 1400 requests for one account history of 151
transactions). I implemented a dictionary to store the ISIN codes and go
to the investment details page only once per investment, thus reducing
the amount of requests to 241 for the same history.
The outputs of "history 1" for the same life insurance are identical
before and after this patch.

Closes: 833427@redmine