[boursorama] Update market accounts with the latest balance
The balance on the accounts page does not match the balance on the
invests page, resulting in incoherences between account balance and the
sum of invests valuations.
This patch enables the update of market account balances from the
invests page to get the latest amount and to be coherent with the sum of
investments.

Closes: 6613@zendesk