[creditmutuel] Fixed access to subscriptions & docs with subbank in URL
The subbank item was missing in the URl construction, leading to the
login page instead of the subscriptions page. This was fixed by
specifying the sub bank in the URL.
I also specified the xpath to get all the subscribers since "//option"
was way to loose and led to other unrelated elements in the HTML page.

Closes: 6867@zendesk, 6868@zendesk