[caissedepargne] Prevent Boobill crash for professional accounts
The Boobill implementation was not done for "netpro" accounts (that have
a different website structure) leading to crashes during the "assert".
I added a "raise NotImplementedError" for these accounts.
In addition, some private customers have not activated the e-Ddocuments
yet so we must verify that subscriptions are activated before trying to
fetch them.

Closes : 19746@Sibi