[bforbank] Retry added because document won't load sometimes.
The summary with the data regarding the accounts do not load sometimes
for some reason. It does not apepar to be time-related but rather try-related.
When manually connecting to the customer's account in-browser, I have to manually
reload the page several times (between 1 and 3) before obtaining the right data.
Hence the parameter tries=5, in case it doesn't load.