[hsbc] Add profile capability
This commit adds the profile capability. On "mes informations"->"mettre
a jour mes données" the page contains a div_err div with "Service
indisponible" in it. But we can still see some informations about the
profile: the name and address(without the city...).
On the website, the other informations are loaded after, but nothing is
really usefull: the phone number shows "xxxxx9999"  and the
email shows "xxxxxxxpont@mail.com"