[creditmaritime] Update regions from creditmaritime
Creditmarime is an AbstractModule from Banquepopulaire.
This commit changes the website url because they made a fusion between
'atlantique' and 'bretagnenormandie' regions.
We need to insert a new url: redirect url. We are redirected on this url
when we try to log in and this url changes according to the website we come from:
banquepopulaire and 'creditmaritime sud ouest' have the same, 'creditmaritime
grand ouest' uses an other one.
In the future, it is highly probable that the sudouest will have the same format as
CMGO.