[milleis] Corrected removal of untyped accounts
Several customers complained that their Life Insurances were not
scraped.
A validate() method excluded the accounts that are UNKNOWN.
I removed it and placed a logger warning to catch all the untyped accounts in
order to type them in the future.
I also changed the condition on investments in case there is none and removed the condition
excluding investments without quantity.

Closes: 6810@zendesk, 6834@zendesk, 6978@zendesk, 7184@zendesk