[boursorama] Fixed card matching and card history
Website tags changed and the code did not match the card numbers
anymore.
This was fixed with correct card numbers.
Sometimes several cards match the same bank account so I added code to
match using the name of the card holder. I tested with all the related
connections and it works correctly.
I left the assert in case the account hash AND the card holder name were
not enough to match the card number.

Closes: 16376@Sibi