[creditcooperatif] new espace bourse for Credit Cooperatif's connections
The new Linebourse's website uses new requests to an API that are implemented
in a brand new browser class (LinebourseAPIBrowser).

Caisse d'Epargne's connections still use the old Linebourse website,
that's why an additional class attribute stores the browser to use
(LINEBOURSE_BROWSER).

It is only used by Credit cooperatif's connections so far, so the
existing code has to be kept as such, using if-conditionals on the
current url to choose which requests to run.

Also, raise an ActionNeeded for the first connection (CGUs)