[cragr] Get PEA correct balance (without liquidities)
Some PEA accounts are only present on the market space and not on the
main page, and since we don't update the balance for PEA accounts in
market_accounts_matching(), these PEA simply did not have any balance.

With this patch the PEA accounts are correctly updated by going to the
investments details in order to fetch their balance WITHOUT liquidities
corresponding to the balance of the DAV PEA account.

Closes: 6613@zendesk