[lcl] Added non handled pages for Calie space
A customer with many accounts had a crash when the module tried to get
to the history of a "Calie" account. In fact some Calie pages were not
handled. I also added an assert before trying
self.page.get_detail() because it often happens that we simply are on
the wrong page when doing that (NoneType, DiscPage...)
This patch should fix Unhandled Calie pages.
The assert should catch any unhandled page.

Closes: 7438@zendesk