[caissedepargne] take into account any connection type
There are several kind of Caisse d'Epargne connections. They are
described by a "typeAccount" parameter, according to the fact that the
connection is either a "particulier", "professsionel", "personnes
protégées", "Cenet" or even many of these at the same time.

The commit allows the module to deal with any connection type, by
abstracting the code parts that were hardcoding the connection types.

Moreover, in case there are several connection type, it tests each of
them one by one untile the login completes. This is because we can't
find ahead of time which connection type to use when only one of them
can be used to log in.